split ops into submodules
