+++ properties/scene.py
@@ -694,27 +694,30 @@ class AppleseedRenderSettings(bpy.types.PropertyGroup):
     override_mode: bpy.props.EnumProperty(name="override_mode",
                                           items=[
                                               ('albedo', "Albedo", ""),
-                                              ('facing_ratio', "Facing Ratio", ""),
-                                              ('coverage', "Coverage", ""),
-                                              ('geometric_normal', "Geometric Normal", ""),
                                               ('ambient_occlusion', "Ambient Occlusion", ""),
+                                              ('assemblies', "Assemblies", ""),
                                               ('assembly_instances', "Assembly Instances", ""),
                                               ('barycentric', "Barycentric", ""),
                                               ('bitangent', "Bitangent", ""),
+                                              ('coverage', "Coverage", ""),
                                               ('depth', "Depth", ""),
+                                              ('facing_ratio', "Facing Ratio", ""),
+                                              ('geometric_normal', "Geometric Normal", ""),
                                               ('materials', "Materials", ""),
+                                              ("objects", "Objects", ""),
                                               ('object_instances', "Object Instances", ""),
                                               ('original_shading_normal', "Original Shading Normal", ""),
                                               ('primitives', "Primitives", ""),
                                               ('ray_spread', "Ray Spread", ""),
-                                              ('regions', "Regions", ""),
                                               ('screen_space_wireframe', "Screen Space Wireframe", ""),
+                                              ('screen_space_velocity', "Screen Space Velocity", ""),
                                               ('shading_normal', "Shading Normal", ""),
                                               ('sides', "Sides", ""),
                                               ('tangent', "Tangent", ""),
                                               ('uv', "UVs", ""),
                                               ('world_space_position', "World Space Position", ""),
-                                              ('world_space_wireframe', "World Space Wireframe", "")],
+                                              ('world_space_wireframe', "World Space Wireframe", ""),
+                                              ('world_space_velocity', "World Space Velocity", "")],
                                           default='facing_ratio')
