+++ ui/materials.py
@@ -66,7 +66,7 @@ class ASMATERIAL_PT_slots(bpy.types.Panel):
             col.operator("object.material_slot_add", icon='ADD', text="")
             col.operator("object.material_slot_remove", icon='REMOVE', text="")
 
-            col.menu("MATERIAL_MT_specials", icon='DOWNARROW_HLT', text="")
+            # col.menu("MATERIAL_MT_specials", icon='DOWNARROW_HLT', text="")
 
             if is_sortable:
                 col.separator()
